🎨 CommandHandler 클래스를 삭제하고 여러 함수들로 만든다
